add BaseConfiguration_v2 class for v2, v2b configurations to inherit from
