+++ larpix/configs/chip/default_v2b.json
@@ -101,12 +101,6 @@
         "v_cm_lvds_tx1": 5,
         "v_cm_lvds_tx2": 5,
         "v_cm_lvds_tx3": 5
-
-
-
-        
-
-
     }
 }
