+++ larpix/configuration/__init__.py
@@ -1,5 +1,6 @@
 from .configuration import *
 from .configuration_v1 import *
+from .configuration_v2_base import *
 from .configuration_v2 import *
 from .configuration_lightpix_v1 import *
 from .configuration_v2b import *
